Fixed annealing edge cases and acceptance accounting

+++ annealing.hpp
@@ -1,6 +1,6 @@
 #include "utilities.hpp"
 
 double temperatureDecrease(double temperature);
-unsigned int stepIncrease(unsigned int temproundsteps);
+unsigned long stepIncrease(unsigned long temproundsteps);
 double energyDecreaseProbDistr(double energyDifference, double temperature);
 bool acceptreject(double probability, gsl_rng* RNG);
